JOB APPLICATION/requirements.txt
streamlit
python-dotenv
groq
python-docx
PyPDF2
requests
paddleocr
paddlepaddle
